Insist on UTF-8 paths everywhere
As the very helpful README for the `camino` crate taught me, the thing
where file *contents* need to refer to filenames is called the "Makefile
problem." We have that here! It was taking a *lot* of boring code to
preserve `OsString` tidiness w/r/t all paths; now we just give up on
that and require everything to be well-formed UTF-8. So we can intermix
strings and filenames much more freely, which is **really** helpful for
all the annoying details involved in Ninja emission.