Don't embed data files
This would be nice for distribution, but it's very inconvenient to
require generating auxiliary files *during build-script emission*. It
makes `--mode emit` pollute the working directory, for example. Let's
just keep them as normal files and refer to them this way!

The alternative I can think of is to have them emitted *at build time*
with a special rule:

    rule get-file
      command = fud2 --get-file $in > $out

Could work!